requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
django-allauth==65.8.0
django-appconf==1.1.0
django-extensions==4.1
django-imagekit==5.0.0
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pilkit==3.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
